backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone-client
pydantic
langchain
langchain-openai
langchain-community
pandas
numpy
scikit-learn
boto3
requests

# install command: pip install -r requirements.txt